+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * Upgrade to fluvio-future 0.4.0 ([#2470](https://github.com/infinyon/fluvio/pull/2470))
 * Add support to detecting smartmodule type from WASM payload on SPU  ([#2457](https://github.com/infinyon/fluvio/issues/2457))
 * Require `version` field in connector yaml. ([#2472](https://github.com/infinyon/fluvio/pull/2472))
+* Add support to SmartModule parameters in Connectors ([#2455](https://github.com/infinyon/fluvio/issues/2455))
+* Move smartmodule configuration to top level config ([#2471](https://github.com/infinyon/fluvio/issues/2471))
 
 ## Platform Version 0.9.30 - 2022-06-29
 * Improve CLI error output when log_dir isn't writable ([#2425](https://github.com/infinyon/fluvio/pull/2425))